cli/requirements.txt
attrs==23.1.0
azure-core==1.29.4
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
idna==3.4
isodate==0.6.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lark==1.1.7
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
pyproject_hooks==1.0.0
python-hcl2==4.3.2
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
